+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # ChangeLog
 
-### Unreleased
+## 06 May 2025
 
+### malai 0.2.3
+
+- Implemented `malai tcp` and `malai tcp-bridge` to expose any TCP service over
+  kulfi network.
 - Implemented `malai folder`. You can now share a folder with people without
   having to manually run another HTTP server. Requires `--public` flag as no
   ACL yet, also readonly mode for now.
-- Implemented `malai tcp` and `malai tcp-bridge` to expose any TCP service over
-  kulfi network.
 
 ## 30 April 2025